+++ types-derive-internals/src/elastic_type/mod.rs
@@ -73,14 +73,14 @@ fn impl_elastic_type(crate_root: Tokens, item: &syn::MacroInput, mapping: &syn::
     let ty = &item.ident;
 
     quote!(
-        impl #crate_root::document::DocumentType<#mapping> for #ty { }
+        impl #crate_root::derive::DocumentType<#mapping> for #ty { }
     )
 }
 
 // Implement DocumentMapping for the mapping
 fn impl_object_mapping(crate_root: Tokens, mapping: &syn::Ident, es_ty: &syn::Lit) -> Tokens {
     quote!(
-        impl #crate_root::document::prelude::DocumentMapping for #mapping {
+        impl #crate_root::derive::DocumentMapping for #mapping {
             fn name() -> &'static str { #es_ty }
         }
     )
@@ -95,7 +95,7 @@ fn impl_props_mapping(crate_root: Tokens,
     let stmts = prop_ser_stmts;
 
     quote!(
-        impl #crate_root::document::mapping::PropertiesMapping for #mapping {
+        impl #crate_root::derive::PropertiesMapping for #mapping {
             fn props_len() -> usize { #stmts_len }
 
             fn serialize_props<S>(state: &mut S) -> ::std::result::Result<(), S::Error> 
@@ -115,9 +115,9 @@ fn get_props_ser_stmts(crate_root: Tokens, fields: &[(syn::Ident, &syn::Field)])
             let lit = syn::Lit::Str(name.as_ref().to_string(), syn::StrStyle::Cooked);
             let ty = &field.ty;
 
-            let expr = quote!(#crate_root::document::mapping::<#ty, _, _>());
+            let expr = quote!(#crate_root::derive::mapping::<#ty, _, _>());
 
-            quote!(try!(#crate_root::document::field_ser(state, #lit, #expr));)
+            quote!(try!(#crate_root::derive::field_ser(state, #lit, #expr));)
         })
         .collect();
