+++ types/src/date/format.rs
@@ -1,6 +1,6 @@
 use chrono;
 use chrono::{DateTime, UTC};
-use chrono::format::{Parsed, Item, DelayedFormat};
+use chrono::format::{Item, DelayedFormat};
 use std::error::Error;
 use std::fmt::{Display, Result as FmtResult, Formatter};
 use std::vec::IntoIter;
@@ -61,30 +61,6 @@ pub trait DateFormat
     ///
     /// This is the string used when defining the format in the field mapping.
     fn name() -> &'static str;
-}
-
-/// Parse a date string using an owned slice of items.
-pub fn parse_from_tokens<'a>(date: &str, fmt: Vec<Item<'a>>) -> Result<DateTime<UTC>, ParseError> {
-    let mut parsed = Parsed::new();
-    match chrono::format::parse(&mut parsed, date, fmt.into_iter()) {
-        Ok(_) => {
-            // If the parsed result doesn't contain any time, set it to the default
-            if parsed.hour_mod_12.is_none() {
-                let _ = parsed.set_hour(0);
-                let _ = parsed.set_minute(0);
-            }
-
-            // Set the DateTime result
-            let dt = try!(parsed.to_naive_datetime_with_offset(0));
-            Ok(chrono::DateTime::from_utc(dt, chrono::UTC))
-        }
-        Err(e) => Err(e.into()),
-    }
-}
-
-/// Format a date string using an owned slice of items.
-pub fn format_with_tokens<'a>(date: &DateTime<UTC>, fmt: Vec<Item<'a>>) -> FormattedDate<'a> {
-    date.format_with_items(fmt.into_iter()).into()
 }
 
 enum FormattedDateInner<'a> {
